Update repository interface

+++ Repositories/Interfaces/ICustomerRepository.cs
@@ -1,11 +1,10 @@
 ﻿using BusinessObjects.Models;
-using System.Reflection.Metadata;
 
 namespace Repositories.Interfaces
 {
     public interface ICustomerRepository
     {
         Task<IReadOnlyList<Customer>> GetAllCustomersAsync(CancellationToken ct = default);
-        Task<Customer?> GetCustomerByIdAsync(int id, CancellationToken ct = default);
+        Task<Customer?> GetCustomerByIdAsync(long id, CancellationToken ct = default);
     }
 }
