+++ Repositories/Interfaces/IMenuCategoryRepository.cs
@@ -6,6 +6,6 @@ namespace Repositories.Interfaces
     public interface IMenuCategoryRepository
     {
         Task<IReadOnlyList<MenuCategory>> GetAllMenuCategoriesAsync(CancellationToken ct = default);
-        Task<MenuCategory?> GetMenuCategoryByIdAsync(int id, CancellationToken ct = default);
+        Task<MenuCategory?> GetMenuCategoryByIdAsync(long id, CancellationToken ct = default);
     }
 }
